refactor(reader): extract ChapterMapOverlay for reuse across EPUB and CBZ readers
